Adding support for stripWWW and stripFragments
With this pull requests users of this library can pass in option object
and utilize strip www and strip fragments from the normalize-url
library.

This is done by allowing users to pass in optional option parameter as
follows

    getUrls(text, {stripFragment: false});

and

    getUrls(text, {stripWWW: false});

This change is backwards compatible, where the defaults favor the
functionality as before.

Note that The normalize-url library was updated to 1.3.0.